+++ src/main/java/com/opus/opus/modules/contest/api/ContestController.java
@@ -18,7 +18,6 @@ import com.opus.opus.modules.contest.application.dto.response.ContestVoteStatist
 import com.opus.opus.modules.contest.application.dto.response.ContestVotesLimitResponse;
 import com.opus.opus.modules.contest.application.dto.response.VotePeriodResponse;
 import com.opus.opus.modules.member.domain.Member;
-import com.opus.opus.modules.team.application.TeamQueryService;
 import com.opus.opus.modules.team.application.dto.ImageResponse;
 import com.opus.opus.modules.team.application.dto.response.MemberVoteCountResponse;
 import com.opus.opus.modules.contest.application.dto.response.ContestRankingResponse;
@@ -50,7 +49,6 @@ public class ContestController {
 
     private final ContestCommandService contestCommandService;
     private final ContestQueryService contestQueryService;
-    private final TeamQueryService teamQueryService;
 
     @GetMapping("/{contestId}/image/banner")
     public ResponseEntity<Resource> getContestBanner(@PathVariable final Long contestId) {
@@ -172,28 +170,28 @@ public class ContestController {
     @GetMapping("/{contestId}/votes/me")
     public ResponseEntity<MemberVoteCountResponse> getMemberVoteCount(@PathVariable Long contestId,
                                                                       @LoginMember Member member) {
-        final MemberVoteCountResponse response = teamQueryService.getMemberVoteCount(member.getId(), contestId);
+        final MemberVoteCountResponse response = contestQueryService.getMemberVoteCount(member.getId(), contestId);
         return ResponseEntity.ok(response);
     }
 
     @Secured("ROLE_관리자")
     @GetMapping("/{contestId}/ranking")
     public ResponseEntity<List<ContestRankingResponse>> getTeamRanking(@PathVariable final Long contestId) {
-        final List<ContestRankingResponse> responses = teamQueryService.getTeamRanking(contestId);
+        final List<ContestRankingResponse> responses = contestQueryService.getTeamRanking(contestId);
         return ResponseEntity.ok(responses);
     }
 
     @Secured("ROLE_관리자")
     @GetMapping("/{contestId}/votes/statistics")
     public ResponseEntity<ContestVoteStatisticsResponse> getVoteStatistics(@PathVariable final Long contestId) {
-        final ContestVoteStatisticsResponse response = teamQueryService.getVoteStatistics(contestId);
+        final ContestVoteStatisticsResponse response = contestQueryService.getVoteStatistics(contestId);
         return ResponseEntity.ok(response);
     }
 
     @Secured("ROLE_관리자")
     @GetMapping("/{contestId}/submissions")
     public ResponseEntity<List<ContestSubmissionResponse>> getTeamSubmissions(@PathVariable final Long contestId) {
-        final List<ContestSubmissionResponse> responses = teamQueryService.getTeamSubmissions(contestId);
+        final List<ContestSubmissionResponse> responses = contestQueryService.getTeamSubmissions(contestId);
         return ResponseEntity.ok(responses);
     }
 }
