+++ src/main/java/com/opus/opus/modules/team/domain/dao/TeamSubmissionResult.java
@@ -5,6 +5,6 @@ public record TeamSubmissionResult(
         String teamName,
         String projectName,
         String trackName,
-        boolean isSubmitted
+        Boolean isSubmitted
 ) {
 }
